+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/display/MenuButton.kt
@@ -41,5 +41,12 @@ internal class MenuButton(
 
     fun dismissMenu() = impl.dismissMenu()
 
+    /**
+     * Sets a lambda to be invoked when the menu is dismissed
+     */
+    fun setMenuDismissAction(onDismiss: () -> Unit) {
+        impl.onDismiss = onDismiss
+    }
+
     fun setColorFilter(@ColorInt color: Int) = impl.setColorFilter(color)
 }